+++ flutter/lib/src/sentry_flutter.dart
@@ -7,7 +7,9 @@ import 'package:meta/meta.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import '../sentry_flutter.dart';
 import 'event_processor/android_platform_exception_event_processor.dart';
+import 'integrations/screenshot_integration.dart';
 import 'native_scope_observer.dart';
+import 'renderer/renderer.dart';
 import 'sentry_native.dart';
 import 'sentry_native_channel.dart';
 
@@ -32,11 +34,15 @@ mixin SentryFlutter {
     @internal PackageLoader packageLoader = _loadPackageInfo,
     @internal MethodChannel channel = _channel,
     @internal PlatformChecker? platformChecker,
+    @internal RendererWrapper? rendererWrapper,
   }) async {
     final flutterOptions = SentryFlutterOptions();
 
     if (platformChecker != null) {
       flutterOptions.platformChecker = platformChecker;
+    }
+    if (rendererWrapper != null) {
+      flutterOptions.rendererWrapper = rendererWrapper;
     }
 
     final nativeChannel = SentryNativeChannel(channel, flutterOptions);
@@ -139,6 +145,11 @@ mixin SentryFlutter {
         !platformChecker.isWeb &&
         (platform.isAndroid || platform.isIOS || platform.isMacOS)) {
       integrations.add(LoadImageListIntegration(channel));
+    }
+    final renderer = options.rendererWrapper.getRenderer();
+    if (renderer == FlutterRenderer.skia ||
+        renderer == FlutterRenderer.canvasKit) {
+      integrations.add(ScreenshotIntegration());
     }
 
     integrations.add(DebugPrintIntegration());